+++ changelog.md
@@ -1,6 +1,6 @@
 # v1.0.2
 - Fix casing in `waitForBadgified`
-- Fix crash when using DevTools' `RenderNode`
+- Fix crash when using DevTools' `RenderNode` <c-33333b>(i hope)</c>
 
 # v1.0.1
 - Fix comment colors of Elder Moderators not appearing properly.